fix: only create WebSocket on component mount
With no dependencies, this useEffect() hook was creating a new WebSocket
connection on every re-render of the page, which led to unbounded growth
as the dashboard is open all Hack Night. Changing this to only create
the WebSocket on component mount should fix the issue. The connection
will reconnect automatically if disconnected because of the
reconnecting-websocket library.